+++ test/android_os_test.dart
@@ -2,10 +2,20 @@
 
 import 'package:flutter_test/flutter_test.dart';
 
-//import 'package:flutter_android/android_os.dart';
+import 'package:flutter_android/android_os.dart';
 
 void main() {
   group('android_os', () {
-    // TODO
+    group('Parcel', () {
+      test(".obtain()", () {
+        expect(Parcel.obtain(), isNot(equals(null)));
+      });
+      test("#writeValue()", () {
+        // TODO
+      });
+    });
+    group('Parcelable', () {
+      // TODO
+    });
   });
 }